+++ c_code/sim.cpp
@@ -41,7 +41,7 @@ int main(int argc, char* argv[]){
         ("s,size", "grid size", cxxopts::value<int>()->default_value("500"))
         ("specrate", "speciation rate", cxxopts::value<double>()->default_value("0.0001"))
         ("r,reps", "number of repetitions", cxxopts::value<int>()->default_value("1"))
-        ("m,mutsize", "maximum change in mutation event", cxxopts::value<int>()->default_value("0.1"))
+        ("m,mutsize", "maximum change in mutation event", cxxopts::value<double>()->default_value("0.1"))
         ("o,outfile", "output file name", cxxopts::value<std::string>()->default_value("out.csv"))
         ("h,help", "Print usage")
         ;
@@ -55,7 +55,7 @@ int main(int argc, char* argv[]){
     outfile = result["outfile"].as<std::string>();
     size = result["size"].as<int>();
     nrep = result["reps"].as<int>();
-    mutsize = result["mutsize"].as<int>();
+    mutsize = result["mutsize"].as<double>();
     specrate = result["specrate"].as<double>();
 
     timescale = 100*(size*1.0/p);
